% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{auto}
\alias{auto}
\title{Automobile data from Stata}
\format{
A data frame with 74 rows and 12 variables
\describe{
  \item{make}{Make and Model}
  \item{price}{Price}
  \item{mpg}{Mileage (mpg)}
  \item{rep78}{Repair Record 1978}
  \item{headroom}{Headroom (in.)}
  \item{trunk}{Trunk space (cu. ft.)}
  \item{weight}{Weight (lbs.)}
  \item{length}{Length (in.)}
  \item{turn}{Turn Circle (ft.)}
  \item{displacement}{Displacement (cu. in.)}
  \item{gear_ratio}{Gear Ratio}
  \item{foreign}{Car type; 0 = Domestic, 1 = Foreign}
}
}
\source{
Chambers, J. M., W. S. Cleveland, B. Kleiner, and P. A. Tukey. 1983. Graphical Methods for Data Analysis. Belmont, CA: Wadsworth.
}
\usage{
auto
}
\description{
This data, which comes standard in Stata, originally came from the April 1979 issue of Consumer Reports and from the United States Government EPA statistics on fuel consumption; they were compiled and published by Chambers et al. (1983).
}
\details{
This data is used in the \emph{Probability and Regression Review} chapter of \emph{Causal Inference: The Mixtape}.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
