% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{avocado}
\alias{avocado}
\title{Data on avocado sales}
\format{
A data frame with 169 rows and 3 variables:
\describe{
  \item{Date}{Date of observation}
  \item{AveragePrice}{Average avocado price}
  \item{TotalVolume}{Total volume of avocados sold}
}
}
\source{
Kiggins, Justin. 2018. \url{https://www.kaggle.com/neuromusic/avocado-prices/}
}
\usage{
avocado
}
\description{
This data set includes information on the average price and total amount of avocados sold across 169 weeks from 2015 to 2018. This data covers only sales of 'conventional' avocados that take place in California.
}
\details{
This data was used in the \emph{Identification} chapter of The Effect by Huntington-Klein
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
