% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{castle}
\alias{castle}
\title{Data on castle-doctrine statutes and violent crime}
\format{
A data frame with 19584 rows and 21 variables
\describe{
\item{year}{Year}
\item{sid}{state id}
\item{robbery_gun_r}{Region-quarter fixed effects}
\item{jhcitizen_c}{justifiable homicide by private citizen count}
\item{jhpolice_c}{justifiable homicide by police count}
\item{homicide}{homicide count per 100,000 state population}
\item{robbery}{Region-quarter fixed effects}
\item{assault}{aggravated assault count per 100,000 state population}
\item{burglary}{burglary count per 100,000 state population}
\item{larceny}{larceny count per 100,000 state population}
\item{motor}{motor vehicle theft count per 100,000 state population}
\item{murder}{murder count per 100,000 state population}
\item{unemployrt}{unemployment rate}
\item{blackm_15_24}{\% of black male aged 15-24}
\item{whitem_15_24}{\% of white male aged 15-24}
\item{blackm_25_44}{\% of black male aged 25-44}
\item{whitem_25_44}{\% of white male aged 25-44}
\item{poverty}{poverty rate}
\item{r20001-r20104}{Region-quarter fixed effects}
\item{trend_1-trend_51}{State linear time trends}
\item{l_homicide}{Logged crime rate}
\item{l_larceny}{Logged crime rate}
\item{l_motor}{Logged crime rate}
\item{l_police}{Logged police presence}
\item{l_income}{Logged income}
\item{l_prisoner}{Logged number of prisoners}
\item{l_lagprisoner}{Lagged log prisoners}
\item{l_exp_subsidy}{Logged subsidy spending}
\item{l_exp_pubwelfare}{Logged public welfare spending}
\item{lead1-lag5}{Indicators of how many time periods until/since treatment}
\item{popwt}{Population weight}
}
}
\source{
Cheng, Cheng, and Mark Hoekstra. 2013. “Does Strengthening Self-Defense Law Deter Crime or Escalate Violence? Evidence from Expansions to Castle Doctrine.” Journal of Human Resources 48 (3): 821–54.
}
\usage{
castle
}
\description{
This data looks at the impact of castle-doctrine statutes on violent crime. Data from the FBI Uniform Crime Reports Summary files are combined with information on castle-doctrine/stand-your-ground law impementation in different states.
}
\details{
This data is used in the \emph{Difference-in-Differences} chapter of \emph{Causal Inference: The Mixtape} by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
