% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ccdrug}
\alias{ccdrug}
\title{Data on Drug Arrests from the Crown Court Sentencing Survey}
\format{
A data frame with 16973 rows and 45 variables
\describe{
  \item{custody}{Taken in to custody.}
  \item{male}{Is a male}
  \item{first_offense}{This is the first offense}
  \item{age}{Age in ten-year bins}
  \item{offense}{Offense type}
  \item{prev_convictions}{Previous convictions, in bins of None, 1-3, 4-9, or 10+}
  \item{drg_class}{Type of drug}
  \item{drg_culpability}{Level of culpability for crime}
  \item{drg_increasing_x, drg_reducing_x}{A set of indicators that should increase or reduce the likelihood of being taken into custody. See variable labels for specific definitions.}
}
}
\source{
Pina Sanchez, J., & Harris, L., 2020. Sentencing gender? Investigating the presence of gender disparities in Crown Court sentences. \emph{Criminal Law Review}, 2020(1), pp. 3-28.
}
\usage{
ccdrug
}
\description{
The \code{ccdrug} data contains data on drug arrests in England and Wales from 2012 through 2015.
}
\details{
This data set is used in the \emph{Partial Identification} chapter of \emph{The Effect}.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
