% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{close_college}
\alias{close_college}
\title{Data from Card (1995) to estimate the effect of college education on earnings}
\format{
A data frame with 3010 rows and 8 variables
\describe{
\item{lwage}{Log wages}
\item{educ}{Years of education}
\item{exper}{Years of work experience}
\item{black}{Race: Black}
\item{south}{In the southern United States}
\item{married}{Is married}
\item{smsa}{In a Standard Metropolitan Statistical Area (urban)}
\item{nearc4}{There is a four-year college in the county}
}
}
\source{
Card, David. 1995. “Aspects of Labour Economics: Essays in Honour of John Vanderkamp.” In. University of Toronto Press.
}
\usage{
close_college
}
\description{
Data from the National Longitudinal Survey Young Men Cohort. This data is used to estimate the effect of college education on earnings, using the presence of a nearby (in-county) college as an instrument for college attendance.
}
\details{
This data is used in the \emph{Instrumental Variables} chapter of \emph{Causal Inference: The Mixtape} by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
