% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{credit_cards}
\alias{credit_cards}
\title{Data on Taiwanese Credit Card Holders}
\format{
A data frame with 30000 rows and 4 variables
\describe{
\item{LateSept}{Credit card payment is late in Sept 2005}
\item{LateApril}{Credit card payment is late in April 2005}
\item{BillApril}{Total bill in April 2005 in thousands of New Taiwan Dollars}
\item{AGE}{Age of card-holder}
}
}
\source{
Lichman, Moshe. 2013. UCI Machine Learning Repository.
}
\usage{
credit_cards
}
\description{
Data from the UCI Machine Learning Repository on Taiwanese credit card holders, the amount of their credit card bill, and whether their payment was late.
}
\details{
This data is used in the \emph{Matching} chapter of \emph{The Effect} by Huntington-Klein.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
