% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gapminder}
\alias{gapminder}
\title{Gapminder data}
\format{
A data frame with 1704 rows and 6 variables
\describe{
  \item{country}{The country}
  \item{continent}{The continent the country is in}
  \item{year}{The year data was collected. Ranges from 1952 to 2007 in increments of 5 years}
  \item{lifeExp}{Life expectancy at birth, in years}
  \item{pop}{Population}
  \item{gdpPercap}{GDP per capita (US$, inflation-adjusted)}
}
}
\source{
\url{https://www.gapminder.org/data/}

Jennifer Bryan (2017). gapminder: Data from Gapminder. R package version 0.3.0. \url{https://CRAN.R-project.org/package=gapminder}
}
\usage{
gapminder
}
\description{
The \code{gapminder} data contains data on life expectancy and GDP per capita by country and year.
}
\details{
This data set is the same one found in the \emph{gapminder} package in R as of 2020. This data set is used in the \emph{Fixed Effects} chapter of \emph{The Effect}.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
