% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gov_transfers_density}
\alias{gov_transfers_density}
\title{Data from "Government Transfers and Political Support" for Density Tests}
\format{
A data frame with 52549 rows and 1 variable:
\describe{
  \item{Income_Centered}{Income measure, centered around program cutoff (negative value = eligible)}
}
}
\source{
Manacorda, M., Miguel, E. and Vigorito, A., 2011. Government transfers and political support. \emph{American Economic Journal: Applied Economics}, 3(3), pp.1-28.
}
\usage{
gov_transfers_density
}
\description{
The \code{gov_transfers_density} data contains data from Manacorda, Miguel, and Vigorito (2011) on government transfer program that was administered based on an income cutoff. As opposed to the \code{gov_transfers} data set, this data set only contains income information, but has a wider range of it, for use with density discontinuity tests.
}
\details{
This data is used in the \emph{Regression Discontinuity} chapter of \emph{The Effect}.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
