% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{mortgages}
\alias{mortgages}
\title{Data from "How do Mortgage Subsidies Affect Home Ownership? Evidence from the Mid-Century GI Bills"}
\format{
A data frame with 214144 rows and 6 variables
\describe{
  \item{bpl}{Birth State}
  \item{qob}{Quarter of birth}
  \item{nonwhite}{White/nonwhite race indicator. 1 = Nonwhite}
  \item{vet_wwko}{Veteran of either the Korean war or World War II}
  \item{home_ownership}{Owns a home}
  \item{qob_minus_kw}{Quarter of birth centered on eligibility for mortgage subsidy (0+ = eligible)}
}
}
\source{
Fetter, D.K., 2013. How do mortgage subsidies affect home ownership? Evidence from the mid-century GI bills. \emph{American Economic Journal: Economic Policy}, 5(2), pp.111-47.
}
\usage{
mortgages
}
\description{
The \code{mortgages} data contains data from Fetter (2015) on home ownership rates by men, focusing on whether they were born at the right time to be eligible for mortgage subsidies based on their military service.
}
\details{
This data is used in the \emph{Regression Discontinuity} chapter of \emph{The Effect}.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
