% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhefs_codebook}
\alias{nhefs_codebook}
\title{NHEFS Codebook}
\format{
A data frame with 64 rows and 2 variables. Each row contains the name
  of a variable (`variable`) in [`nhefs`] and [`nhefs_complete`] and a
  description of what the variable represents (`description`).
}
\source{
[wwwn.cdc.gov/nchs/nhanes/nhefs/](wwwn.cdc.gov/nchs/nhanes/nhefs/)
}
\usage{
nhefs_codebook
}
\description{
`nhefs_codebook` is the codebook for [`nhefs`] and [`nhefs_complete`].
}
\keyword{datasets}
