% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhefs_complete}
\alias{nhefs_complete}
\title{National Health and Nutrition Examination Survey Data I Epidemiologic Follow-up Study}
\format{
A data frame with 1556 rows and 67 variables. The codebook is available as `nhefs_codebook`.
}
\source{
[wwwn.cdc.gov/nchs/nhanes/nhefs/](wwwn.cdc.gov/nchs/nhanes/nhefs/)
}
\usage{
nhefs_complete
}
\description{
`nhefs_complete` is the same as [`nhefs`], but only participants
with complete data are included. The variables that need to be complete to be
included are: `qsmk`, `sex`, `race`, `age`, `school`, `smokeintensity`,
`smokeyrs`, `exercise`, `active`, `wt71`, `wt82`, and `wt82_71`.
}
\keyword{datasets}
