% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{organ_donations}
\alias{organ_donations}
\title{Organ Donation Data}
\format{
A data frame with 162 rows and 3 variables
\describe{
  \item{State}{The state, where California is the Treated group}
  \item{Quarter}{Quarter of observation, in "Q"QYYYY format}
  \item{Rate}{Organ donation rate}
  \item{Quarter_Num}{Quarter of observation in numerical format. 1 = Quarter 4, 2010}
}
}
\source{
Kessler, J.B. and Roth, A.E., 2014. Don't take 'no' for an answer: An experiment with actual organ donor registrations. National Bureau of Economic Research working paper No. 20378. \url{https://www.nber.org/papers/w20378}
}
\usage{
organ_donations
}
\description{
The \code{organ_donation} data contains data from Kessler and Roth (2014) on organ donation rates by state and quarter. The state of California enacted an active-choice phrasing for their organ donation sign-up questoin in Q32011. The only states included in the data are California and those that can serve as valid controls; see Kessler and Roth (2014).
}
\details{
This data is used in the \emph{Difference-in-Differences} chapter of \emph{The Effect}.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
