% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{restaurant_inspections}
\alias{restaurant_inspections}
\title{Data on Restaurant Inspections}
\format{
A data frame with 27178 rows and 5 variables
\describe{
  \item{business_name}{Name of restaurant/chain}
  \item{inspection_score}{Health Inspection Score}
  \item{Year}{Year of inspection}
  \item{NumberofLocations}{Number of locations in restaurant chain}
  \item{Weekend}{Was the inspection performed on a weekend?}
}
}
\source{
Camus, Louis-Ashley. 2020. \url{https://www.kaggle.com/loulouashley/inspection-score-restaurant-inspection}
}
\usage{
restaurant_inspections
}
\description{
The \code{restaurant_inspections} data contains data on restaurant health inspections performed in Anchorage, Alaska.
}
\details{
This data set is used in the \emph{Regression} chapter of \emph{The Effect}.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
