% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ri}
\alias{ri}
\title{A simple simulated data set for calculating p-values}
\format{
A data frame with 8 rows and 5 variables
\describe{
  \item{name}{Fictional Name}
  \item{d}{Treatment}
  \item{y}{Outcome}
  \item{y0}{Outcome if untreated}
  \item{y1}{Outcome if treated}
}
}
\source{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\usage{
ri
}
\description{
This simulated data allows for a quick and easy calculation of a p-value using randomization inference.
}
\details{
This data is used in the \emph{Potential Outcomes Causal Model} chapter of \emph{Causal Inference: The Mixtape} by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
