% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{snow}
\alias{snow}
\title{Data from John Snow's 1855 study of the cause of cholera}
\format{
A data frame with 4 rows and 4 variables
\describe{
\item{year}{Year}
\item{supplier}{Water pump supplier}
\item{treatment}{Status of water pump}
\item{deathrate}{Deaths per 10k 1851 population}
}
}
\source{
Snow, John. 1855. 'On the Mode of Communication of Cholera'. John Churchill."

Coleman, Thomas. 2019. 'Causality in the time of cholera: John Snow as a prototype for causal inference.' SSRN 3262234."
}
\usage{
snow
}
\description{
A subset of the aggregated death rate data from Snow's legendary study of the source of the London Cholera outbreak.
}
\details{
This data is used in the \emph{Difference-in-Differences} chapter of \emph{The Effect} by Huntington-Klein.
}
\references{
Huntington-Klein. 2021. The Effect: An Introduction to Research Design and Causality. \url{https://theeffectbook.net}.
}
\keyword{datasets}
