% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{thornton_hiv}
\alias{thornton_hiv}
\title{Data from HIV information experiment in Thornton (2008)}
\format{
A data frame with 4820 rows and 7 variables
\describe{
\item{villnum}{Village ID}
\item{got}{Got HIV results}
\item{distvct}{Distance in kilometers}
\item{tinc}{Total incentive}
\item{any}{Received any incentive}
\item{age}{Age}
\item{hiv2004}{HIV results}
}
}
\source{
Thornton, Rebecca L. 2008. 'The Demand for, and Impact of, Learning Hiv Status.' American Economic Review 98 (5): 1829–63.
}
\usage{
thornton_hiv
}
\description{
\code{thornton_hiv} comes from an experiment in Malawi looking at whether cash incentives could encourage people to learn the results of their HIV tests.
}
\details{
This data is used in the Potential Outcomes Causal Model chapter of Causal Inference: The Mixtape by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
