% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{titanic}
\alias{titanic}
\title{Data from the sinking of the Titanic}
\format{
A data frame with 4820 rows and 7 variables
\describe{
\item{villnum}{Village ID}
\item{got}{Got HIV results}
\item{distvct}{Distance in kilometers}
\item{tinc}{Total incentive}
\item{any}{Received any incentive}
\item{age}{Age}
\item{hiv2004}{HIV results}
}
}
\source{
British Board of Trade (1990), Report on the Loss of the ‘Titanic’ (S.S.). British Board of Trade Inquiry Report (reprint). Gloucester, UK: Allan Sutton Publishing.
}
\usage{
titanic
}
\description{
\code{titanic} comes from the sinking of the Titanic, and can be used to look at survival by different demographic characteristics.
}
\details{
This data is used in the Matching and Subclassification chapter of Causal Inference: The Mixtape by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
