% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{training_bias_reduction}
\alias{training_bias_reduction}
\title{Simulated data from a job training program for a bias reduction method}
\format{
A data frame with 8 rows and 4 variables
\describe{
\item{Unit}{Unit ID}
\item{Y}{Outcome}
\item{D}{Treatment}
\item{X}{Matching variable}
}
}
\source{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\usage{
training_bias_reduction
}
\description{
This simulated data is used to demonstrate the bias-reduction method in matching as per Abadie and Imbens (2011).
}
\details{
This data is used in the \emph{Matching and Subclassification} chapter of \emph{Causal Inference: The Mixtape} by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
