% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{training_example}
\alias{training_example}
\title{Simulated data from a job training program}
\format{
A data frame with 25 rows and 9 variables
\describe{
\item{unit_treat}{Unit ID for treated observations}
\item{age_treat}{age for treated observations}
\item{earnings_treat}{earnings for treated observations}
\item{unit_control}{Unit ID for control observations}
\item{age_control}{age for control observations}
\item{earnings_control}{earnings for control observations}
\item{unit_matched}{Unit ID for matched controls}
\item{age_matched}{age for matched controls}
\item{earnings_matched}{earnings for matched controls}
}
}
\source{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\usage{
training_example
}
\description{
This simulated data, which is presented in the form of a full results, table, is used to demonstrate a matching procedure.
}
\details{
This data is used in the \emph{Matching and Subclassification} chapter of \emph{Causal Inference: The Mixtape} by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
