% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{yule}
\alias{yule}
\title{Data on 19th century English Poverty from Yule (1899)}
\format{
A data frame with 32 rows and 5 variables
\describe{
  \item{location}{Location in England}
  \item{paup}{Pauperism Growth}
  \item{outrelief}{Poverty Relief Growth}
  \item{old}{Annual growth in aged population}
  \item{pop}{Annual growth in population}
}
}
\source{
Yule, G. Udny. 1899. 'An Investigation into the Causes of Changes in Pauperism in England, Chiefly During the Last Two Interensal Decades.' Journal of Royal Statistical Society 62: 249–95.
}
\usage{
yule
}
\description{
\code{yule} allows for a look at the correlation between poverty relief and poverty rates in England in the 19th century.
}
\details{
This data is used in the Potential Outcomes Causal Model chapter of Causal Inference: The Mixtape by Cunningham.
}
\references{
Cunningham. 2021. Causal Inference: The Mixtape. Yale Press. \url{https://mixtape.scunning.com/index.html}.
}
\keyword{datasets}
