% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causaldata-package.R
\docType{package}
\name{causaldata-package}
\alias{causaldata}
\alias{causaldata-package}
\title{causaldata: Example Data Sets for Causal Inference Textbooks}
\description{
This package contains example data sets to run the example
    problems from causal inference textbooks. Currently, contains data
    sets for The Effect by Nick Huntington-Klein, Causal Inference: The
    Mixtape by Scott Cunningham, and Causal Inference: What If? By Miguel
    Hernán and James Robins.
}
\author{
\strong{Maintainer}: Nick Huntington-Klein \email{nhuntington-klein@seattleu.edu} (\href{https://orcid.org/0000-0002-7352-3991}{ORCID})

Authors:
\itemize{
  \item Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})
}

}
\keyword{internal}
