% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{greek_data}
\alias{greek_data}
\title{Data from a fictional randomized heart transplant study}
\format{
A data frame with 20 rows and 4 variables:
\describe{
\item{name}{The name of a Greek god}
\item{l}{A prognostic factor}
\item{a}{The treatment, a heart transplant}
\item{y}{The outcome, death}
}
}
\source{
Hernán and Robins. Causal Inference. \url{https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book/}
}
\usage{
greek_data
}
\description{
\code{greek_data} is a fictional data set from Table 2.2 in Chapter 2 of Causal Inference. From the book: "Table 2.2 shows the data from our heart transplant randomized study. Besides data on treatment A (1 if the individual received a transplant, 0 otherwise) and outcome Y (1 if the individual died, 0 otherwise), Table 2.2 also contains data on the prognostic factor L (1 if the individual was in critical condition, 0 otherwise), which we measured before treatment was assigned."
}
\keyword{datasets}
