% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhefs}
\alias{nhefs}
\title{National Health and Nutrition Examination Survey Data I Epidemiologic Follow-up Study}
\format{
A data frame with 1629 rows and 67 variables. The codebook is available as \code{nhefs_codebook}.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/nhefs/}
}
\usage{
nhefs
}
\description{
\code{nhefs} is a cleaned data set of the data used in Causal Inference by Hernán and Robins. \code{nhefs} is dataset containing data from the National Health and Nutrition Examination Survey Data I Epidemiologic Follow-up Study (NHEFS). The NHEFS was jointly initiated by the National Center for Health Statistics and the National Institute on Aging in collaboration with other agencies of the United States Public Health Service. A detailed description of the NHEFS, together with publicly available data sets and documentation, can be found at \url{https://wwwn.cdc.gov/nchs/nhanes/nhefs/}.
}
\references{
Hernán and Robins. Causal Inference. \url{https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book/}
}
\keyword{datasets}
