% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{nhefs_codebook}
\alias{nhefs_codebook}
\title{NHEFS Codebook}
\format{
A data frame with 64 rows and 2 variables.
\describe{
\item{variable}{The variable being described}
\item{description}{The variable description}
}
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/nhefs/}
}
\usage{
nhefs_codebook
}
\description{
\code{nhefs_codebook} is the codebook for \code{nhefs} and \code{nhefs_complete}.
}
\references{
Hernán and Robins. Causal Inference. \url{https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book/}
}
\keyword{datasets}
