% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhefs_complete}
\alias{nhefs_complete}
\title{Complete-Data National Health and Nutrition Examination Survey Data I Epidemiologic Follow-up Study}
\format{
A data frame with 1556 rows and 67 variables. The codebook is available as \code{nhefs_codebook}.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/nhefs/}
}
\usage{
nhefs_complete
}
\description{
\code{nhefs_complete} is the same as \code{nhefs}, but only participants with complete data are included. The variables that need to be complete to be included are: \code{qsmk}, \code{sex}, \code{race}, \code{age}, \code{school}, \code{smokeintensity}, \code{smokeyrs}, \code{exercise}, \code{active}, \code{wt71}, \code{wt82}, and \code{wt82_71}.
}
\references{
Hernán and Robins. Causal Inference. \url{https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book/}
}
\keyword{datasets}
